Add tests for scripts
