+++ .github/workflows/release.yml
@@ -35,13 +35,13 @@ jobs:
         env:
           UV_PUBLISH_TOKEN: ${{ secrets.pypi_token }}
         run: |
-          ./release.sh
+          scripts/release.sh
       - name: Build Changelog
         id: changelog
         if: env.release_tag
         run: |
           echo 'CHANGELOG_BODY<<EOF' >> $GITHUB_ENV
-          uv run dev_tools.py get-changelog ${{ env.release_tag }} >> $GITHUB_ENV
+          uv run scripts/dev_tools.py get-changelog ${{ env.release_tag }} >> $GITHUB_ENV
           echo 'EOF' >> $GITHUB_ENV
       - name: Export config schema
         if: env.release_tag
