+++ scripts/tests/conftest.py
@@ -1,0 +1,37 @@
+import os
+from http import client
+
+import pytest
+from vcr import VCR
+from vcr.stubs import VCRHTTPConnection, VCRHTTPSConnection
+
+VCR_CASSETTE_DIR = os.path.join(os.path.dirname(__file__), 'cassettes')
+VCR_RECORD_MODE = os.environ.get('VCR_RECORD_MODE', 'once')
+
+vcr = VCR(
+    cassette_library_dir=VCR_CASSETTE_DIR,
+    record_mode=VCR_RECORD_MODE,
+    custom_patches=(
+        (client, 'HTTPSConnection', VCRHTTPSConnection),
+        (client, 'HTTPConnection', VCRHTTPConnection),
+    ),
+)
+
+
+@pytest.fixture
+def online(request, monkeypatch):
+    """Be applied automatically to any test using the `online` mark.
+
+    It will record and playback network sessions using VCR.
+
+    The record mode of VCR can be set using the VCR_RECORD_MODE environment variable when running tests.
+    """
+    if VCR_RECORD_MODE == 'off':
+        yield None
+    else:
+        module = request.module.__name__.split('tests.')[-1]
+        class_name = request.cls.__name__
+        cassette_name = f'{module}.{class_name}.{request.function.__name__}'
+        cassette_path = os.path.join(VCR_CASSETTE_DIR, cassette_name)
+        with vcr.use_cassette(path=cassette_path) as cassette:
+            yield cassette
