+++ scripts/tests/update_changelog/test_1/ChangeLog.md
@@ -1,0 +1,23 @@
+---
+title: ChangeLog
+description: 
+published: true
+date: 2024-03-15T00:33:34.805Z
+tags: 
+editor: markdown
+dateCreated: 2022-09-18T04:48:39.193Z
+---
+
+# Changelog
+This changelog is in progress. It can be manually updated via the wiki, but is also updated automatically via select commit messages and new releases. The two comment lines with git hashes (`<!---a1234--->`) must not be changed or removed.
+
+<!---85bfdefbd3d2cd4ab258f1111dbc1ad074a0fcd3--->
+
+## 3.13.19.dev (unreleased)
+<!---659bf21e5852415df1463c124ef67b30189b4a37--->
+
+## 3.13.18 (2025-01-20)
+[all commits](https://github.com/Flexget/Flexget/compare/v3.13.17...v3.13.18)
+
+## 3.13.17 (2025-01-19)
+[all commits](https://github.com/Flexget/Flexget/compare/v3.13.16...v3.13.17)
