+++ scripts/tests/update_changelog/test_2/ChangeLog.md
@@ -1,0 +1,32 @@
+---
+title: ChangeLog
+description: 
+published: true
+date: 2024-03-15T00:33:34.805Z
+tags: 
+editor: markdown
+dateCreated: 2022-09-18T04:48:39.193Z
+---
+
+# Changelog
+This changelog is in progress. It can be manually updated via the wiki, but is also updated automatically via select commit messages and new releases. The two comment lines with git hashes (`<!---a1234--->`) must not be changed or removed.
+
+<!---345282dafb789006d27653f28d651003dcf19012--->
+
+## 3.13.19.dev (unreleased)
+### Fixed
+- Fix a bug
+
+### Added
+- Add a new feature
+
+### Deprecated
+- A plugin is now deprecated ([#1522](https://github.com/Flexget/Flexget/issues/1522))
+
+<!---659bf21e5852415df1463c124ef67b30189b4a37--->
+
+## 3.13.18 (2025-01-20)
+[all commits](https://github.com/Flexget/Flexget/compare/v3.13.17...v3.13.18)
+
+## 3.13.17 (2025-01-19)
+[all commits](https://github.com/Flexget/Flexget/compare/v3.13.16...v3.13.17)
